Release ready: Moved api keys to blinkupconfig.js file.
Goal: We don't want users to merge in their API keys in the sample application.
Solution: Created a .example file which demonstrate how to build a blinkupconfig.js file, which is git ignored.
This file is generated during an AS / Xcode build, with empty strings, but only if it doesn't exist.
Users then get to update the root www/js/blinkupconfig.js file with their api keys. Risk of committing it is lowered.
Updated readme accordingly.

Tested:
- File isn't overwritten when it already exists.
- API Key is used by blinkup.